backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic[email]==2.5.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.28.0

# Vector Database
weaviate-client==3.25.1

# AI/ML
flask
flask-cors
transformers
torch
accelerate
peft
bitsandbytes
safetensors

openai==1.3.5
langchain==0.0.350
sentence-transformers==2.2.2
transformers>=4.31.0
torch>=2.0.0
accelerate>=0.21.0
peft>=0.4.0
# bitsandbytes is not supported on Windows (CUDA/GPUs on Linux only). Install only on Linux:
bitsandbytes>=0.41.0; platform_system != "Windows"
safetensors>=0.3.1

# Windows notes:
# - On Windows CPU setups, torch will install a CPU build (OK). For CUDA, install a matching wheel from https://pytorch.org/get-started/locally/
# - If you see errors about BitsAndBytesConfig, it will be gracefully skipped in code when bitsandbytes is unavailable.

# Cache
redis==5.0.1

# Utils
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-json-logger==2.0.7

# NOTE: StudentGPT runtime deps (transformers/torch/accelerate/peft/bitsandbytes/safetensors)
# are included above in AI/ML section. `backend/studentgpt/requirements.txt` is deprecated.

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
black==23.11.0
isort==5.12.0
mypy==1.7.0
types-requests==2.31.0.10
types-python-dateutil==2.8.19.14
